// src/modules/nugecid/domain/enums/status-desarquivamento.enum.ts
export enum StatusDesarquivamentoEnum {
  FINALIZADO = "FINALIZADO",
  DESARQUIVADO = "DESARQUIVADO",
  NAO_COLETADO = "NAO_COLETADO",
  SOLICITADO = "SOLICITADO",
  REARQUIVAMENTO_SOLICITADO = "REARQUIVAMENTO_SOLICITADO",
  RETIRADO_PELO_SETOR = "RETIRADO_PELO_SETOR",
  NAO_LOCALIZADO = "NAO_LOCALIZADO",
}